Extract resource ordering to new submodule
