+++ localstack-core/localstack/services/cloudformation/provider.py
@@ -89,10 +89,12 @@ from localstack.services.cloudformation.engine.entities import (
     StackSet,
 )
 from localstack.services.cloudformation.engine.parameters import strip_parameter_type
-from localstack.services.cloudformation.engine.template_deployer import (
+from localstack.services.cloudformation.engine.resource_ordering import (
     NoResourceInStack,
-    NoStackUpdates,
     order_resources,
+)
+from localstack.services.cloudformation.engine.template_deployer import (
+    NoStackUpdates,
 )
 from localstack.services.cloudformation.engine.template_utils import resolve_stack_conditions
 from localstack.services.cloudformation.engine.transformers import (
